[style] team: styles 파일 분리
