+++ src/app/(team)/team/_components/TaskListBarList/ProcessBadge.tsx
@@ -1,4 +1,4 @@
-import { processBadgeWrapperStyle } from '@/app/(team)/team/_components/styles';
+import { processBadgeWrapperStyle } from '@/app/(team)/team/_components/TaskListBarList/styles';
 
 const ProcessBadge = () => {
   return <div className={`${processBadgeWrapperStyle}`}></div>;
