+++ src/app/(team)/team/_components/TeamBanner/styles.ts
@@ -1,0 +1,24 @@
+import clsx from 'clsx';
+import {
+  teamItemWrapperStyle,
+  RADIUS,
+} from '@/app/(team)/team/_components/styles';
+
+// 📌TeamBanner.tsx style
+export const teamBannerWrapperStyle = clsx(
+  'relative flex items-center justify-between p-6',
+  teamItemWrapperStyle,
+  'h-[64px] w-full',
+  'border border-slate-50/10 bg-[#272e3f]',
+  RADIUS
+);
+
+export const teamBannerImgStyle =
+  'tablet:left-[75%] absolute left-1/2 -translate-x-1/2';
+
+export const teamBannerTitleStyle = clsx(
+  'text-xl-bold',
+  'relative flex-1',
+  'laptop:max-w-[800px] tablet:max-w-[460px] max-w-[245px]',
+  'min-w-0'
+);
